fix: main.py dont start and rabbitmq unification
