+++ requirements.txt
@@ -2,3 +2,4 @@ fastapi==0.135.2
 uvicorn==0.35.0
 pymongo==4.13.2
 pydantic==2.11.7
+pika>=1.3.0
