+++ src/core/constants.py
@@ -7,3 +7,4 @@ GENESIS_TRANSACTION_TYPE = "GENESIS_ISSUANCE"
 CHAIN_METADATA_ID = "chain_state"
 BLOCKCHAIN_EVENTS_EXCHANGE = "ufrocoin.blockchain.events"
 GENESIS_EVENT_ROUTING_KEY = "genesis.created"
+TRANSACTION_EVENT_ROUTING_KEY = "transaction.created"
