+++ src/main.py
@@ -1,6 +1,7 @@
 from fastapi import FastAPI
-feature/valid-block-structure
 from src.api.block_router import router as block_router
+from src.api.global_router import router as global_router
+from src.api.transaction_router import router as transaction_router
 from src.core.database import close_db_client
 
 app = FastAPI(
@@ -9,9 +10,14 @@ app = FastAPI(
     version="1.0.0",
 )
 
+app.include_router(global_router)
 app.include_router(block_router, prefix="/api")
+app.include_router(transaction_router, prefix="/api")
 
 
 @app.on_event("shutdown")
 def shutdown_event() -> None:
     close_db_client()
+
+
+
